Refactor check_deployment_status to remove the need for the deployment info file
We can use other methods to determine what state we are in.

Instead of using the existance of the deployment file, we can use
the schema version being 0 to tell us that we are a new deployment.

Instead of using the server PV data we can use the existence of
our server record to determine if we are a new replica.

Instead of comparing a version string we can just do an upgrade if
we have pending migrations. This will work for named and numbered
versions.