Remove references to the .deployment_info file and region PV
We no longer need the .deployment_info file as it was just being
used to track version numbers. This is no longer needed after the
change to the check_deployment_status function.

This file was the last one being stored on the region PV so it
also can be removed