Add images and alter template for containerized workers
Now we have a base image which contains the manageiq code base
then we build the orchestrator image and worker images from there.

The only image that contains the comiled assets is the UI image.

Every image that runs a web server exposes only port 3000, so we
make apache listen on that port and the UI worker binds to port 3001

A common liveness check script (manageiq_liveness_check) is used for
all images which don't have a service.

Each web server worker has it's own service which is pointed to
by httpd directly. This means that we can also remove the
"manageiq" service.