Revert to DeploymentConfig and remove server PV
This makes the entire app act as one server. The GUID is generated
using a template parameter and is provided to the server pod using
an environment variable.

We ignore logs because we will be printing our logs to STDOUT which
will be processed by the EFK log aggregation stack.

We also remove certs because they will be handled by the httpd
pod or by the route itself.

This allows us to remove the "server" PV entirely.