Add the ManageIQ copr repo to the app image
We need access to this repo to install the http-parser package
which was removed from EPEL when RHEL 7.4 was released.

It will be available again when CentOS 7.4 is released, but that
won't be for at least a few more weeks.

See https://bugzilla.redhat.com/show_bug.cgi?id=1481470 for more
details