Use httpd service account for all environments
This allows us to use the same template on minishift as well
as a full OpenShift cluster. Now the change that needs to be
made based on environment is just in documentation steps
rather than editing the template.

Technically, this always uses the miq-httpd service account for
the httpd pod and allows the user to choose the scc assigned to
that service account based on their environment. For minishift
that scc will be miq-sysadmin and for the full cluster, it will
be anyuid.