Make websockets work again
This commit changes the proxy directives added in #208
to rewrite rules which seem to play better with the existing
rules put in place for auth.

Before this, notifications did not work properly and httpd processes
were starting constantly and not exiting.

https://bugzilla.redhat.com/show_bug.cgi?id=1526601

Note:
This seems to have been working before 7e86fe90, but somehow
the RewriteRule added there messes with the ws proxy.