Consolidate orchestrator startup scripts
Many of the functions used by the orchestrator entrypoint script
were only used by the orchestrator, but were defined in the
base image.

This commit moves the functions or the content of the functions
directly into the orchestrator entrypoint script and does the
same for scripts directly in the orchestrator image which were
unnecessarily split out.

This makes the orchestrator entrypoint more readable and maintainable