Move the database info into its own secret
This allows us to set the database to whatever we want when deploying
the secret then only deploying the database deployment if we need it.
This eliminates the need for the duplicate external database template.

Additionally to do this we need to build up the DATABASE_URL environment
variable in the container rather than in the template. To do this
a new file was added which, when sourced into the current session,
will provide the required environment variables.