Separate secrets template into manageiq app secrets and pg secrets
This will improve the user experience of a deploy script by not
requiring a user to re-specify a generated value if they only want
to specify one.

For example, we don't want to re-generate the encryption key if the
user was just trying to change their database password. Additionally
we wouldn't want to force the user to go fetch the encryption key
string to just provide it to the template so the existing value
wouldn't be changed.