Move set -e under the deployment status
Previously this would exit the script after evm:deployment_status
returned because we use the (non-zero) return code to determine
which status we are in.

Additionally add a log line to say we are checking the status
otherwise it looks like it's taking us a ton of time to write
the encryption key.