Add information about ingress/routes to README
Fixes #92

Creating the ingress like this requires a hostname and an ssl
cert and key at every deployment. Previously, it was created for
the user by OpenShift and thus would never match the application
domain out of the box.

Now a user can pre-create the secret used by the ingress with the
desired hostname before deployment and have the certs be correct
the first time through.