Add a hooks url for build-time customization
The idea here is that anything that needs to be overlayed onto
the source directories will be handled by this one script.
It should assume that it's running locally to the "build",
"app", "sui", and "appliance" directories and add whatever it
wants to those, knowing that they will be copied over to the final
image in the next stage.

Fixes #360