Fix the backup and restore jobs
This moves the backup and restore jobs to templates and alters
the pod spec so that they consume the new secret format correctly.

Use template parameter generation to create a unique name for each
backup and restore job.

Fixes #285

Additionally this gets us a bit closer to solving #287 as we can
now provide the specific backup we want to restore as a template
parameter.