Add the actual steps for using the operator to the README
This puts the steps to build and run the operator in the README
instead of referencing the OpenShift documentation.

Additionally it reorders some of the sections to ensure that the
deployment works properly. Specifically, the CRD needs to be created
before the operator is deployed to the cluster. This also needs to be
done whether you are running locally or deploying the operator as
a pod in the cluster