Move the postgres secrets out of the CR
Now a user must provide a secret name when deploying the CR
We will either use the existing secret or will create a secret with
the given name containing default and generated values.

This also adds support for generating a database password if one
was not provided and moves the port into the database secret for
both the operator and template cases.