Add a field to the manageiq spec for the tls secret name
This will behave just like the postgres secret:

If it is not provided a new secret called "tls-secret" will be created.
That secret will be removed when the CR is removed.

If the name is provided, but the secret does not exist when the
app is deployed a new secret will be created with the provided
name and used. This secret will also be removed with the CR.

If the name of an existing secret is provided, it will be used
and will not be removed when the CR is removed.

Fixes #409