Define and use a role for the orchestrator
This is more like how an operator works and allows us to control
the individual permissions given to the orchestrator.

Specifically the "edit" cluster role didn't include the ability to
set finalizers on pods. This was preventing us from using kubernetes
garbage collection to ensure that the worker deployments were cleaned
up whenever the orchestrator pod is brought down.

Part of #428