Encode httpd and postgresql conf as static strings in functions
Previously this only worked if you ran the operator-sdk locally
because the files were not copied into the image and the working
directory at runtime would have been different.
This will ensure that the string is available to us at runtime
in the container as well as running locally.

Fixes #453