Don't run the app as root
A bunch of certifications require us not to run as root
in containers. Additionally the vast majority of the application
should run fine without root and the parts that need root are
typically things we can't do in containers anyway.

Specifically this changes the permissions of some directories to
allow the root group access which is what we'll be running as when
using the restricted scc in OpenShift