Deploy an httpd image without init if it's not needed
We don't need an init system unless we're using an external auth
method that requires integration with sssd and dbus.

This means that we can deploy without elevated permissions when we're
using internal (database) authentication or oidc (which doesn't require
sssd and dbus).

This commit adds the logic to switch between the two deployments
depending on the authentication type which is now a top-level
parameter in the CR.

https://github.com/ManageIQ/manageiq-pods/issues/467