Configure httpd env and volumes conditionally based on auth type
The environment variables are only needed for oidc, the
httpd-auth-config volume is needed for other external auth methods
and "internal" auth doesn't need any of that. This ensures that
we only deploy what is really needed.

Additionally this will create a new volume for user-provided conf
when HttpdAuthConfig is present and HttpdAuthenticationType is not
internal

Fixes #467