Reverse the logic for the local login parameter
This also changes the boolean parameters to be pointers to properly
express the concept of "unset". Without this change the ManageIQSpec
struct would have them set to `false` by default because that is
bool's zero value. This made it impossible to tell if a user was
explicitly setting a value to false or if it was unset.