Remove extra sha256 sum when generating the encryption key
This was present in the [original implementation]
(https://github.com/ManageIQ/manageiq-password/blob/f85e888fe60433390d0a1d5145889edc67437fdf/lib/manageiq/password.rb#L210-L213)
but was only needed because that code allowed for user-provided
passwords to be used as the encryption key. In our implementation
we are only using random bytes as the input so an extra sha sum
doesn't do anything to improve security.

Additionally this was triggering an issue in some static security
analysis tools.