Add a CR parameter for an initial group name
When provided, we will create a group with the given name with
the super administrator role. The main use case would be when
a user wants to deploy with basic auth totally disabled as soon
as the application comes up. This means that they need at least one
group in the application that matches a group in their identity
provider.

We also now run a full seed after migrating the database on initial
deploy so that we can create the initial group correctly.

Fixes #508