Use controllerutil.CreateOrUpdate to maintain memcached resources
Prior to this change, any changes to the on-cluster CR would not be
reflected in the memcached service and deployment. Additionally
changes to the service and deployment would not be resolved to the
correct state by the operator.

Both of those cases are now handled by this patch.

Applies to #438