Add a call to SetControllerReference for the PG config map
This also removes the generic config map mutate function in favor
of extracting the only piece of shared conditional logic that it should
contain.

The new function (addAppLabel) should be usable by every mutate
function we create and will handle adding the application label
while leaving the rest of the labels untouched.

Addresses one third of #554