Append our pull secret to the SA pull secret list
OpenShift creates a default pull secret with the service account
so before this change we were fighting with OpenShift by removing
their pull secret and replacing it with ours.

This commit checks to see if there are any other secrets in the SA
and appends ours to the list if it isn't already there.