Prefer recreate as the deployment strategy for many services.
Rolling deployments require that a new pod be available before the old pod
is killed for availability reasons.  However we have many services that
should only ever have a single replica and having a second pod can actually
cause some strange bugs.  This went unnoticed for many of our services
since nothing would prevent a second pod from coming up, but it was found
on the postgresql deployment since it needs to mount the data PVC and with
most backing storages are ReadWriteOnce and the pod can not mount the PVC
if another pod has it mounted.

I switched postgresql, kafka, zookeeper, memcached and the orchestrator to
Recreate since there should only ever be one replica of those.  I didn't see
a problem with having multiple replicas of httpd, so I left that as Rolling.