Attempt to pull GUID from existing orchestrator if one exists
In kasparov the source of truth for the server GUID was moved to the CR,
but we didn't have a way to migrate the existing value to the CR.  Now, if
the CR value is empty, we first look for an existing orchestrator pod and
if found, we pull the GUID from the ENV var there, otherwise we generate a
new one.  As before, if a value is specified in the CR, that value is used.