We only need the ruby environment, not full container_env
7d841daecccaacbbe made this work but container_env does expensive URL
escaping of the DATABASE_USER and PASSWORD to build DATABASE_URL which
is not needed in the liveness check. Since all pods run this liveness check
many times per minutes, it has to be as fast as possible.

This drops a few 100ms from the liveness check depending on the environment.