Expand LimitRequestFieldSize in the internal httpd
LimitRequestFieldSize: Expand this to a large number to allow pass-through.
This does not introduce a potential DoS, because the value is validated by
the httpd container first.  However, if a user changes this value in the
httpd container, we need to be able to accomodate that value here also.

512K was chosen as the size because that is also the maximum limit of
the OIDCCacheShmEntrySizeMax value for OIDC.