Remove ovirt-4.4-dependencies.repo
The ovirt-4.4-dependencies.repo brings in a lot of repos, some of which
have old RPMs which have security issues, in particular pyyaml 5.1.2.
We only really need the ovirt-4.4.repo in order to bring in
ovirt-ansible-collection, python3-ovirt-engine-sdk4 and
v2v-conversion-host-ansible.  Removing ovirt-4.4-dependencies.repo
allows many dependencies to install from baseos and appstream which
should be more secure.