Drop secret watcher
We already watch for secret changes in our namespace.  The only advantage
that this had was in the case of a change to the httpd auth config secret
we would only reconcile httpd.  I don't think this added complexity is
worth the time difference.  Also, rewriting all of this code to fit the
latest operator-sdk is very complicated.