Install httpd after replacing the release RPM
httpd will still come from the UBI appstream repo due to the exclusion set above.

When the release RPM is swapped, the RPM scripts from that and shadow-utils
reset the contents of /etc/passwd causing the apache user to be removed and
httpd fails to start due to the missing user.

Based on changes in https://github.com/ManageIQ/manageiq-pods/pull/956