Fix bug in TLS secret detection
If someone were to successfully install the application then delete the
TLSSecret, the operator would not recreate it because it detected that the
Database Secret already existed.  This will cause errors during reconcile
when the operator tries to update the CR status.  Instead, we should be
checking the existence of the TLS Secret.