Call out the default group delimiter for mellon is a semicolon
MellonMergeEnvVars On brings back all groups for REMOTE_USER_GROUPS
By default, the delimiter is ;, but this can be overridden

This is a no-op, but it highlights that a semicolon is used to join groups.
Liberty (via oidc) is using a comma, but that is not displayed here.
LookupUserGroups specifies a ":"