We also need to watch certain secrets that we don't own for changes.
In this case we need to set the resourceVersion of the secret containing
the SSL certs on the pods that use them, but we don't "own" it, so changes
to it weren't triggering a reconcile.  We don't want to "own" this
secret because it would get garbage collected if our CR disappeared
(we don't want that to happen).  So, we can watch all secrets and trigger
a reconcile on any manageiqs that have it as the value of Spec.InternalCertificatesSecret

CP4AIOPS-5110