Watch Ingresses and Routes if available
Allow changes to our Ingress and Route to trigger a reconcile.
Routes are not available outside of OpenShift and the controller setup
will fail if we try to watch resources that are not registered to the
cluster.  So, we first need to see if we get an error if we try to list
routes.

CP4AIOPS-6311