If we are using Routes, ensure that the ingress is removed
Sometimes when we are initially created we go down the ingress path rather
than the routes.  Then certificates are created and we switch to routes
instead.  Unfortunately nothing was cleaning up the ingress (and potentially
associated route httpd-xxxxx if we are running in OpenShift.)  This ensures
that if we are using routes, we clean up any ingress that we may have created
in the past.

CP4AIOPS-6311