Fix issue with a shared reference for the deploymentLabels
This fix handles an issue where the spec.selector.matchLabels and the
spec.template.metadata.labels contained a reference to the same map.
Some labels only are supposed to appear in the template labels and not
the selector, such as the `manageiq.org/backup` label and the
`ssl-certificate-resource-version` label. These labels were accidentally
being copied into the spec selector as well.

In Kubernetes, template labels are modifiable while selector labels are
immutable. Another rule is that template labels must match the selector
labels, but it's a subset (i.e. you can have extra template labels).
When the extra labels are "shared" into the spec selector, they will
initially not exhibit any problems because they match, but as soon as
the value is changed in the template labels, such as updating the
`ssl-certificate-resource-version` to a new version on a cert update,
then you get an error like the following:

```
Deployment.apps "postgresql" is invalid: spec.template.metadata.labels: Invalid value: map[string]string
{"app":"manageiq-application", "manageiq.org/backup":"t", "name":"postgresql", "ssl-certificate-resource-version":"12271329"}
: `selector` does not match template `labels`
```

CP4AIOPS-8731