Fix issue where IPBlock and NamespaceSelector are mutually exclusive
On upgrade, the IPBlock was already set and then after upgrade it enters
the conditional where it tries to add the NamespaceSelector. So, in that
case both are set. Instead, when we set one, we should unset the other
to ensure they aren't both set.

This fixes the error:

    NetworkPolicy.networking.k8s.io "miq-application-allow-inbound-httpd" is
    invalid: spec.ingress[0].from[0]: Forbidden: may not specify both ipBlock
    and another peer

CP4AIOPS-22012